+++ cypress/integration/forms/selects.spec.ts
@@ -5,6 +5,13 @@ describe('Selects', () => {
     cy.get('.page-loader').should('not.exist', { timeout: 20000 });
   });
 
+  const shouldBeNotFocused = () =>
+    cy.get('.ngx-select-input-underline .underline-fill').should('have.css', 'width', '0px');
+  const shouldBeFocused = () =>
+    cy.get('.ngx-select-input-underline .underline-fill').should('not.have.css', 'width', '0px');
+  const shouldBeNotActive = () => cy.root().should('not.have.class', 'active');
+  const shouldBeActive = () => cy.root().should('have.class', 'active');
+
   it('have no detectable a11y violations on load', () => {
     cy.get('ngx-select-input').withinEach($el => {
       cy.checkA11y($el, {
@@ -18,7 +25,10 @@ describe('Selects', () => {
 
   describe('Basic Input', () => {
     beforeEach(() => {
-      cy.get('#select-1').as('CUT');
+      cy.get('ngx-section').first().as('SUT');
+      cy.get('@SUT').find('ngx-select').first().as('CUT');
+      cy.get('@CUT').ngxClose();
+      cy.get('@CUT').clear();
     });
 
     it('has a label', () => {
@@ -35,6 +45,19 @@ describe('Selects', () => {
       cy.get('@CUT').clear().ngxGetValue().should('equal', '');
     });
 
+    it('deselects on click', () => {
+      cy.get('@CUT').ngxGetValue().should('equal', '');
+
+      const text = 'DDOS';
+
+      cy.get('@CUT').select(text).ngxGetValue().should('equal', text);
+
+      cy.get('@CUT').ngxOpen();
+      cy.get('@CUT').find('.ngx-select-dropdown-option').contains(text).click();
+
+      cy.get('@CUT').ngxGetValue().should('equal', '');
+    });
+
     it('selects and clears value twice', () => {
       cy.get('@CUT').ngxGetValue().should('equal', '');
 
@@ -45,11 +68,183 @@ describe('Selects', () => {
       cy.get('@CUT').select(text).ngxGetValue().should('equal', text);
       cy.get('@CUT').clear().ngxGetValue().should('equal', '');
     });
+
+    it('is keyboard accessible', () => {
+      cy.get('@SUT').find('h4').contains('Basic').click();
+
+      cy.get('@CUT').within(() => {
+        // Starts out not focused not open
+        shouldBeNotFocused();
+        shouldBeNotActive();
+
+        // Tab into the select, now focused but still closed
+        cy.realPress('Tab');
+        shouldBeFocused();
+        shouldBeNotActive();
+
+        // Down arrow to open and select first item
+        cy.realPress('ArrowDown');
+        shouldBeFocused();
+        shouldBeActive();
+
+        cy.get('.ngx-select-dropdown-options li li').within(() => {
+          cy.root().first().should('have.class', 'active'); // active
+          cy.root().last().should('not.have.class', 'active'); // not active
+
+          cy.realPress('ArrowDown').realPress('ArrowDown');
+          cy.root().first().should('not.have.class', 'active'); // not active
+          cy.root().last().should('have.class', 'active'); // active
+        });
+
+        // Escape to close list
+        cy.realPress('Escape');
+        shouldBeFocused();
+        shouldBeNotActive();
+
+        // Arrow up to open and select last item
+        cy.realPress('ArrowUp');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(10); // Needed for testing!!!
+
+        cy.get('.ngx-select-dropdown-options li li').within(() => {
+          cy.root().first().should('not.have.class', 'active'); // not active
+          cy.root().last().should('have.class', 'active'); // active
+
+          cy.realPress('ArrowUp');
+          // eslint-disable-next-line cypress/no-unnecessary-waiting
+          cy.wait(10); // Needed for testing!!!
+          cy.realPress('ArrowUp');
+          cy.root().first().should('have.class', 'active'); // active
+          cy.root().last().should('not.have.class', 'active'); // not active
+        });
+
+        // Enter selects an option and closes the list
+        cy.realPress('Enter');
+        shouldBeFocused();
+        shouldBeNotActive();
+        cy.root().ngxGetValue().should('equal', 'Breach');
+
+        // Space opens
+        cy.realPress('Space');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // Space selects an option but leaves list open
+        cy.realPress('ArrowDown');
+        cy.realPress('Space');
+        cy.root().ngxGetValue().should('equal', 'DDOS');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // Can deselect
+        cy.realPress('Space');
+        cy.root().ngxGetValue().should('equal', '');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // Tab away
+        cy.root().ngxClose();
+        cy.realPress('Tab');
+        shouldBeNotFocused();
+        shouldBeNotActive();
+      });
+    });
+  });
+
+  describe('Tagging', () => {
+    beforeEach(() => {
+      cy.get('ngx-section[data-cy=tagging]').as('SUT');
+      cy.get('@SUT').find('ngx-select').first().as('CUT');
+      cy.get('@CUT').ngxClose();
+    });
+
+    it('has a label', () => {
+      cy.get('@CUT').ngxFindLabel().should('contain.text', 'Tagging');
+    });
+
+    it('selects and clears existing values', () => {
+      cy.get('@CUT').ngxGetValue().should('equal', '');
+
+      const text = 'DDOS';
+
+      cy.get('@CUT').select(text).ngxGetValue().should('equal', text);
+
+      // TODO(ngx-ui-testing): tagging should return an array of values
+      cy.get('@CUT').clear().ngxGetValue().should('equal', '');
+    });
+
+    it('can add and clear a new value', () => {
+      cy.get('@CUT').ngxGetValue().should('equal', '');
+
+      const text = 'Other';
+
+      // TODO(ngx-ui-testing): support ngxFill for tagging
+      cy.get('@CUT').find('input').click().type(text).type('{enter}');
+      cy.get('@CUT').ngxGetValue().should('equal', text);
+
+      cy.get('@CUT').clear().ngxGetValue().should('equal', '');
+    });
+
+    it('can add and clear multiple values', () => {
+      cy.get('@CUT').ngxGetValue().should('equal', '');
+
+      cy.get('@CUT').select('DDOS').ngxGetValue().should('equal', 'DDOS');
+      cy.get('@CUT').find('input').click().type('Other').type('{enter}');
+      cy.get('@CUT').ngxGetValue().should('contain', 'DDOS').should('contain', 'Other');
+
+      cy.get('@CUT').find('.ngx-select-clear').first().click();
+      cy.get('@CUT').find('.ngx-select-clear').first().click();
+    });
+
+    it('is keyboard accessible', () => {
+      cy.get('@SUT').find('h4').contains('Basic').click();
+
+      cy.get('@CUT').within(() => {
+        // Starts out not focused not open
+        shouldBeNotFocused();
+        shouldBeNotActive();
+
+        // Tab into the select, now focused and open
+        cy.realPress('Tab');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(100); // Wait for element to refocus
+
+        cy.focused().type('Other{enter}'); // Input should be focused
+
+        cy.root().ngxGetValue().should('contain', 'Other');
+
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(100); // Wait for element to refocus
+
+        // Down arrow to select second item, closes list, stays active
+        cy.realPress('ArrowDown');
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(100); // Needed for testing!!!
+        cy.realPress('Enter');
+
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(120); // Wait for element to refocus on input
+
+        shouldBeFocused();
+        shouldBeActive();
+
+        cy.root().ngxGetValue().should('contain', 'Other').should('contain', 'DDOS');
+
+        // Clears with backspace
+        cy.focused().type('{backspace}{backspace}{backspace}'); // For some reason needs three backspaces in testing
+      });
+    });
   });
 
   describe('Input with AbstractControl', () => {
     beforeEach(() => {
-      cy.get('#select-19').as('CUT');
+      cy.getByName('formCtrl1').as('CUT');
       cy.get('[data-cy=reactiveFormSelectToggleBtn]').as('reactiveFormSelectToggleBtn');
     });
 
@@ -67,7 +262,7 @@ describe('Selects', () => {
 
   describe('Input with AbstractControl disabled by default', () => {
     beforeEach(() => {
-      cy.get('#select-20').as('CUT');
+      cy.getByName('formCtrl2').as('CUT');
     });
 
     it('should be able to be disabled', () => {
@@ -77,7 +272,7 @@ describe('Selects', () => {
 
   describe('Filtering Input', () => {
     beforeEach(() => {
-      cy.get('#select-3').as('CUT');
+      cy.getByName('filtering').as('CUT');
     });
 
     it('has a label', () => {
@@ -95,7 +290,7 @@ describe('Selects', () => {
 
   describe('Templates', () => {
     beforeEach(() => {
-      cy.get('#select-9').as('CUT');
+      cy.get('[data-cy=templates]').as('CUT');
     });
 
     it('selects and clears value', () => {
@@ -111,7 +306,8 @@ describe('Selects', () => {
 
   describe('Multiple Select', () => {
     beforeEach(() => {
-      cy.get('#select-21').as('CUT');
+      cy.get('section header h1').contains('Multi Select').closest('ngx-section').as('SUT');
+      cy.get('@SUT').find('ngx-select').first().as('CUT');
     });
 
     it('selects and clears value', () => {
@@ -121,6 +317,85 @@ describe('Selects', () => {
       cy.get('@CUT').select(['DDOS', 'Physical']).ngxGetValue().should('deep.equal', ['DDOS', 'Physical']);
 
       cy.get('@CUT').clear().ngxGetValue().should('deep.equal', []);
+    });
+
+    it('is keyboard accessible', () => {
+      cy.get('@SUT').find('h4').contains('Basic').click();
+
+      cy.get('@CUT').within(() => {
+        // Starts out not focused not open
+        shouldBeNotFocused();
+        shouldBeNotActive();
+
+        // Tab into the select, now focused but still closed
+        cy.realPress('Tab');
+        shouldBeFocused();
+        shouldBeNotActive();
+
+        // Down arrow to open and select first item
+        cy.realPress('ArrowDown');
+        shouldBeFocused();
+        shouldBeActive();
+
+        cy.get('.ngx-select-dropdown-options li li').within(() => {
+          cy.root().first().should('have.class', 'active'); // active
+          cy.root().last().should('not.have.class', 'active'); // not active
+
+          cy.realPress('ArrowDown').realPress('ArrowDown');
+          cy.root().first().should('not.have.class', 'active'); // not active
+          cy.root().last().should('have.class', 'active'); // active
+        });
+
+        // Escape to close list
+        cy.realPress('Escape');
+        shouldBeFocused();
+        shouldBeNotActive();
+
+        //  Arrow up to open and select last item
+        cy.realPress('ArrowUp');
+        shouldBeFocused();
+        shouldBeActive();
+
+        // eslint-disable-next-line cypress/no-unnecessary-waiting
+        cy.wait(10); // Needed for testing!!!
+
+        cy.get('.ngx-select-dropdown-options li li').within(() => {
+          cy.root().first().should('not.have.class', 'active'); // not active
+          cy.root().last().should('have.class', 'active'); // active
+
+          cy.realPress('ArrowUp');
+          // eslint-disable-next-line cypress/no-unnecessary-waiting
+          cy.wait(10); // Needed for testing!!!
+          cy.realPress('ArrowUp');
+          cy.root().first().should('have.class', 'active'); // active
+          cy.root().last().should('not.have.class', 'active'); // not active
+        });
+
+        // Enter selects an option and leaves list open
+        cy.realPress('Enter');
+        shouldBeFocused();
+        shouldBeActive();
+        cy.root().ngxGetValue().should('deep.equal', ['Breach']);
+
+        // Space selects an option but leaves list open
+        cy.realPress('ArrowDown');
+        cy.realPress('Space');
+        cy.root().ngxGetValue().should('deep.equal', ['Breach', 'DDOS']);
+        shouldBeFocused();
+        shouldBeActive();
+
+        // Can deselect an option
+        cy.realPress('ArrowUp');
+        cy.realPress('Space');
+        cy.root().ngxGetValue().should('deep.equal', ['DDOS']);
+        shouldBeFocused();
+        shouldBeActive();
+
+        cy.root().ngxClose();
+        cy.realPress('Tab');
+        shouldBeNotFocused();
+        shouldBeNotActive();
+      });
     });
   });
 
@@ -177,7 +452,8 @@ describe('Selects', () => {
 
     it('selects value', () => {
       cy.get('@CUT').ngxGetValue().should('deep.equal', '');
-      cy.get('@CUT').ngxFill('dolorem');
+      cy.get('@CUT').ngxOpen();
+      cy.get('@CUT').find('input').ngxFill('dolorem');
       cy.wait('@api');
       cy.get('@CUT').find('li.ngx-select-dropdown-option').should('have.length', 2);
       cy.get('@CUT').select('dolorem eum magni eos aperiam quia');
@@ -191,26 +467,38 @@ describe('Selects', () => {
     });
 
     it('should close on input click', () => {
-      cy.get('@attackType').find('.ngx-select-dropdown-options').should('not.exist');
-      cy.get('@attackType').find('.ngx-select-input-box').click();
-      cy.get('@attackType').find('.ngx-select-dropdown-options').should('be.visible');
+      cy.get('@attackType').within(() => {
+        cy.get('.ngx-select-dropdown-options').should('not.exist');
+        cy.get('.ngx-select-input-box').click();
+        cy.get('.ngx-select-dropdown-options').first().scrollIntoView();
+        cy.get('.ngx-select-dropdown-options').should('be.visible');
+      });
 
-      cy.get('@attackTypeRequired').find('.ngx-select-dropdown-options').should('not.exist');
-      cy.get('@attackTypeRequired').find('.ngx-select-input-box').click();
-      cy.get('@attackTypeRequired').find('.ngx-select-dropdown-options').should('be.visible');
+      cy.get('@attackTypeRequired').within(() => {
+        cy.get('.ngx-select-dropdown-options').should('not.exist');
+        cy.get('.ngx-select-input-box').click();
+        cy.get('.ngx-select-dropdown-options').first().scrollIntoView();
+        cy.get('.ngx-select-dropdown-options').should('be.visible');
+      });
 
       // the current opened select should be closed
       cy.get('@attackType').find('.ngx-select-dropdown-options').should('not.exist');
     });
 
     it('should close on caret down click', () => {
-      cy.get('@attackType').find('.ngx-select-dropdown-options').should('not.exist');
-      cy.get('@attackType').find('.ngx-select-caret').click();
-      cy.get('@attackType').find('.ngx-select-dropdown-options').should('be.visible');
+      cy.get('@attackType').within(() => {
+        cy.get('.ngx-select-dropdown-options').should('not.exist');
+        cy.get('.ngx-select-caret').click();
+        cy.get('.ngx-select-dropdown-options').first().scrollIntoView();
+        cy.get('.ngx-select-dropdown-options').should('be.visible');
+      });
 
-      cy.get('@attackTypeRequired').find('.ngx-select-dropdown-options').should('not.exist');
-      cy.get('@attackTypeRequired').find('.ngx-select-caret').click();
-      cy.get('@attackTypeRequired').find('.ngx-select-dropdown-options').should('be.visible');
+      cy.get('@attackTypeRequired').within(() => {
+        cy.get('.ngx-select-dropdown-options').should('not.exist');
+        cy.get('.ngx-select-caret').click();
+        cy.get('.ngx-select-dropdown-options').first().scrollIntoView();
+        cy.get('.ngx-select-dropdown-options').should('be.visible');
+      });
 
       // the current opened select should be closed
       cy.get('@attackType').find('.ngx-select-dropdown-options').should('not.exist');