+++ projects/swimlane/ngx-ui/src/lib/enums/keyboard-keys.enum.ts
@@ -6,5 +6,7 @@ export enum KeyboardKeys {
   ARROW_DOWN = 'ArrowDown',
   ARROW_UP = 'ArrowUp',
   ARROW_LEFT = 'ArrowLeft',
-  ARROW_RIGHT = 'ArrowRight'
+  ARROW_RIGHT = 'ArrowRight',
+  SPACE = 'Space',
+  TAB = 'Tab'
 }