+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## HEAD (unreleased)
 
+- Enhancement: Improved keyboard accessibility for `ngx-select`
+
 ## 36.1.0 (2021-12-10)
 
 - Enhancement: allow exponential notation for number inputs